Split js.rb into smaller files to ease debugging and readability
Split kv.rb into smaller files
