+++ lib/nats/io/jetstream/js/header.rb
@@ -1,0 +1,31 @@
+# Copyright 2021 The NATS Authors
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+module NATS
+  class JetStream
+    module JS
+      module Header
+        Status              = ("Status".freeze)
+        Desc                = ("Description".freeze)
+        MsgID               = ("Nats-Msg-Id".freeze)
+        ExpectedStream      = ("Nats-Expected-Stream".freeze)
+        ExpectedLastSeq     = ("Nats-Expected-Last-Sequence".freeze)
+        ExpectedLastSubjSeq = ("Nats-Expected-Last-Subject-Sequence".freeze)
+        ExpectedLastMsgID   = ("Nats-Expected-Last-Msg-Id".freeze)
+        LastConsumerSeq     = ("Nats-Last-Consumer".freeze)
+        LastStreamSeq       = ("Nats-Last-Stream".freeze)
+      end
+    end
+  end
+end
