+++ lib/nats/io/msg.rb
@@ -11,7 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-require_relative 'js'
+require_relative 'jetstream'
 
 module NATS
   class Msg
